# swagger/VacinPB.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VacinPB API",
    "description": "Com o caos gerado pela pandemia da COVID-19 e a procura da população pelas vacinas, notamos que em UBS (Unidade Básica de Saúde), os funcionários dessas unidades têm uma enorme dificuldade em fazer a gestão das vacinas, pois precisam fazer as anotações em cadernos para poder realizar todo uma gestão de qual vacina um cidadão tomou, os dados básicos para um cadastro de cada cidadão, o estoque de vacinas, gerando assim, uma demora ainda maior e filas. Visando uma melhora para esses profissionais, desenvolveremos uma API para automatizar todo esse procedimento anterior a aplicação das vacinas.",
    "version": "0.0.1",
    "license": {
      "name": "Lincença: GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "API local"
    }
  ],
  "paths": {
...
    "/vacinas": {
      "post": {
        "summary": "Cadastrar novas vacinas",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vaccine"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 400,
                  "timestamp": "2022-01-18T13:23:29.139+00:00",
                  "fieldErrors": {
                    "field": "campo",
                    "error": "descrição do erro"
                  },
                  "description": "uri=/vacinas"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      },
      "parameters": [
        {
          "name": "lote",
          "in": "query",
          "description": "Filtrar pelo lote da vacina.",
          "required": false,
          "schema": {
            "type": "string"
          },
          "style": "form",
          "explode": true
        },
        {
          "name": "ordenar-validade",
          "in": "query",
          "description": "Filtrar pela validade das vacinas.",
          "required": false,
          "schema": {
            "type": "string"
          },
          "style": "form",
          "explode": true
        }
      ],
      "get": {
        "summary": "Listar todas vacinas",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vaccine"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/vacinas/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer"
          },
          "required": true
        }
      ],
      "get": {
        "summary": "Listar a vacina pelo ID",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vaccine"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "timestamp": "2022-01-18T13:23:29.139+00:00",
                  "message": "recurso não encontrado",
                  "description": "uri=/vacinas/{id}"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar os dados de uma vacina de acordo com o ID informado",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vaccine_id"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 400,
                  "timestamp": "2022-01-18T13:23:29.139+00:00",
                  "fieldErrors": {
                    "field": "campo",
                    "error": "descrição do erro"
                  },
                  "description": "uri=/vacinas/{id}"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "timestamp": "2022-01-18T13:23:29.139+00:00",
                  "message": "recurso não encontrado",
                  "description": "uri=/vacinas/{id}"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      },
      "delete": {
        "summary": "Deletar uma vacina de acordo com o ID informado",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "timestamp": "2022-01-18T13:23:29.139+00:00",
                  "message": "recurso não encontrado",
                  "description": "uri=/vacinas/{id}"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Vaccine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vaccineName": {
            "type": "string",
            "enum": [
              "BCG",
              "HEPATITE B",
              "PENTAVALENTE",
              "POLIOMELITE",
              "PNEUMOCÓCICA",
              "ROTAVÍRUS",
              "MENINGOCÓCICA_C",
              "INFLUENZA",
              "FEBRE_AMARELA",
              "TRIPLICE_VIRAL",
              "DTP",
              "HEPATITE_A",
              "TETRA_VIRAL",
              "VARICELA_ATENUADA",
              "HPV",
              "DT",
              "DTPA",
              "PFIZER",
              "CORONAVAC",
              "JANSSEN",
              "ASTRAZENECA"
            ]
          },
          "manufacturer": {
            "type": "string"
          },
          "manufactureDate": {
            "type": "string",
            "format": "date"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "lotNumber": {
            "type": "string"
          }
        },
        "example": {
          "id": 1,
          "vaccineName": "BCG",
          "manufacturer": "Fundação Ataulpho de Paiva",
          "manufactureDate": "15/11/2021",
          "expirationDate": "15/11/2021",
          "lotNumber": "FXA81L6"
        }
      },
      "Vaccine_id": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vaccineName": {
            "type": "string",
            "enum": [
              "BCG",
              "HEPATITE B",
              "PENTAVALENTE",
              "POLIOMELITE",
              "PNEUMOCÓCICA",
              "ROTAVÍRUS",
              "MENINGOCÓCICA_C",
              "INFLUENZA",
              "FEBRE_AMARELA",
              "TRIPLICE_VIRAL",
              "DTP",
              "HEPATITE_A",
              "TETRA_VIRAL",
              "VARICELA_ATENUADA",
              "HPV",
              "DT",
              "DTPA",
              "PFIZER",
              "CORONAVAC",
              "JANSSEN",
              "ASTRAZENECA"
            ]
          },
          "manufacturer": {
            "type": "string"
          },
          "manufactureDate": {
            "type": "string",
            "format": "date"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "lotNumber": {
            "type": "string"
          }
        },
        "example": {
          "id": 1,
          "vaccineName": "BCG",
          "manufacturer": "Fundação Ataulpho de Paiva",
          "manufactureDate": "15/11/2021",
          "expirationDate": "15/02/2022",
          "lotNumber": "FXA81L6"
        }
      }
    },
    "securitySchemes": {
      "login": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}